Jim closed #123 and reimplemented the fixes upstream. Five conflicts,
all where his version and ours addressed the same defect, and his is
better in every one — so they resolve to his.

The uint[] bind failure is fixed across every database service rather
than only the three we had hit: 53 bigint[] casts against our 3.

Character and server list rows: same conclusion we reached — hold the
list as data and rebuild the rows against the tree that is actually on
screen — but his write-up identifies a case we never diagnosed. Both
panels are authored StartOpen: 0, so on a first login the list arrives
when there is no visual tree at all, and row construction was gated on
the container being non-null. That gate silently dropped every row, and
the player reached an empty character screen with a Create button and no
explanation. We only ever found the later hide/show path.

Row text colour: he uses --abyss-200 where we used --abyss-300, with the
measurements to justify it — 3.36:1 against the panel background, under
the 4.5:1 needed for 12px text, versus 7.01:1. Ours restored visibility
but landed below the accessibility threshold, which we did not check.

He also marks the full-screen login forms CanDrag => false, since a
panel that fills the screen has nowhere to be dragged to.

Kept from this branch, none of it duplicated upstream: the list/select
in-flight gate split, the client auth-token clear tracing, the CLI build
output path, starting equipment and the regen tuning.

FishMMO-DB, ServerAuth and ClientAuth all build with 0 errors.

+++ FishMMO-Database/FishMMO-DB/Npgsql/Services/Scene/Character/CharacterBankService.cs
@@ -155,7 +155,13 @@ namespace FishMMO.Database.Npgsql.Services
 				var id = await ExecuteScalarLongAsync(
 						dbContext,
 						sql,
-						new object[] { item.CharacterID, item.Slot, item.Version, item.TemplateID, item.Seed, item.Amount, now },
+						// Amount is projected to long for the same reason the batched path does it: Npgsql
+						// cannot bind System.UInt32 at all ("The CLR type System.UInt32 isn't natively
+						// supported"), and it throws on the scalar just as it does on the array. The column
+						// is bigint, so long binds exactly. This is the live equip/unequip path —
+						// CharacterInventorySystem calls this overload once per item — so leaving it
+						// unprojected kept every runtime item write failing after the batch path was fixed.
+						new object[] { item.CharacterID, item.Slot, item.Version, item.TemplateID, item.Seed, (long)item.Amount, now },
 						cancellationToken).ConfigureAwait(false);
 
 				if (id <= 0)
